+++ src/main/java/com/lowdragmc/mbd2/integration/create/machine/MBDKineticMachineBlockEntity.java
@@ -12,7 +12,7 @@ import com.simibubi.create.content.kinetics.KineticNetwork;
 import com.simibubi.create.content.kinetics.base.IRotate;
 import com.simibubi.create.content.kinetics.base.KineticBlockEntity;
 import com.simibubi.create.content.kinetics.base.KineticEffectHandler;
-import com.simibubi.create.foundation.utility.Lang;
+import com.simibubi.create.foundation.utility.CreateLang;
 import lombok.Getter;
 import lombok.Setter;
 import net.minecraft.ChatFormatting;
@@ -207,8 +207,8 @@ public class MBDKineticMachineBlockEntity extends KineticBlockEntity implements 
         boolean added = super.addToGoggleTooltip(tooltip, isPlayerSneaking);
         float stressBase = this.calculateAddedStressCapacity();
         if (stressBase != 0.0F && IRotate.StressImpact.isEnabled()) {
-            Lang.translate("gui.goggles.generator_stats").forGoggles(tooltip);
-            Lang.translate("tooltip.capacityProvided").style(ChatFormatting.GRAY).forGoggles(tooltip);
+            CreateLang.translate("gui.goggles.generator_stats").forGoggles(tooltip);
+            CreateLang.translate("tooltip.capacityProvided").style(ChatFormatting.GRAY).forGoggles(tooltip);
             float speed = this.getTheoreticalSpeed();
             if (speed != this.getGeneratedSpeed() && speed != 0.0F) {
                 stressBase *= this.getGeneratedSpeed() / speed;
@@ -216,8 +216,8 @@ public class MBDKineticMachineBlockEntity extends KineticBlockEntity implements 
 
             speed = Math.abs(speed);
             float stressTotal = stressBase * speed;
-            Lang.number(stressTotal).translate("generic.unit.stress").style(ChatFormatting.AQUA).space()
-                    .add(Lang.translate("gui.goggles.at_current_speed").style(ChatFormatting.DARK_GRAY))
+            CreateLang.number(stressTotal).translate("generic.unit.stress").style(ChatFormatting.AQUA).space()
+                    .add(CreateLang.translate("gui.goggles.at_current_speed").style(ChatFormatting.DARK_GRAY))
                     .forGoggles(tooltip, 1);
             added = true;
         }