+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Added supports of setting dynamic machine level.
 * Added supports of modifying geckolib animation speed.
 * Fixed In world preview rendering error.
+* Fixed compatibility with the Create-6.0.0